requirements.txt
py-cord>=2.4.1
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=9.5.0